Move requestDependencies
